week 2 created a dice roller
learning how to use variable, functions name spaces, basic addition. learning to pull functions and variables from different codes scripts.

researched how to use the variables like +,-,*,/,++,--,%
