--- GD12_1133_Lab2.5_Fortuna_Cameron/Lab2 Game Porgramming/Program.cs
@@ -1,16 +1,0 @@
-﻿using Lab2_Game_Porgramming.scirpts;
-
-namespace Lab2_Game_Porgramming
-{
-    internal class Program
-    {
-        static void Main(string[] args)
-        {
-           GameManager manager = new GameManager();
-
-            //manager.PlayGame();
-
-            manager.ProgramStart();
-        }
-    }
-   }
