--- GD12_1133_Lab2.5_Fortuna_Cameron/Lab2 Game Porgramming/scirpts/practice value.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Lab2_Game_Porgramming.scirpts
-{
-    internal class practice_value
-    {
-       
-    }
-}
