+++ Lab2 Final/DiceRoller.cs
@@ -4,9 +4,9 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Lab2_Game_Porgramming.scirpts.scirpt
+namespace Lab2_Final
 {
-    public class DiceRoller
+    internal class DiceRoller
     {
         internal int totalSides = 6;
 
@@ -23,10 +23,10 @@ namespace Lab2_Game_Porgramming.scirpts.scirpt
         //}
 
         public int RandomDiceRoll() //telling function when called to retun a integer
-        { 
+        {
             Random rollInstance = new Random(); //making it so the roll instance has a new random variable each time
 
-            previousResults = rollInstance.Next(1, totalSides +1);// setting up what the previous result will be
+            previousResults = rollInstance.Next(1, totalSides + 1);// setting up what the previous result will be
             return previousResults; // saving the random interger to the previous result 
         }
         //Diceroler is a definition of what a something is, DoceRollerInstance is taking that definition and set a peramtor of what it is.
@@ -38,8 +38,6 @@ namespace Lab2_Game_Porgramming.scirpts.scirpt
         //++ increase variable by one
         //-- decrease variable by one
         //  %  devion with remainder from left to right
-
-
 
     }
 }
