+++ Lab2 Final/Program.cs
@@ -1,0 +1,12 @@
+﻿namespace Lab2_Final
+{
+    internal class Program
+    {
+        static void Main(string[] args)
+        {
+            GameManager manager = new GameManager();
+
+            manager.ProgramStart();
+        }
+    }
+}
